chore: move tests to root
