+++ tsconfig.json
@@ -13,6 +13,5 @@
     "strict": true,
     "target": "es6"
   },
-  "include": ["src"],
-  "exclude": ["src/tests"]
+  "include": ["src"]
 }
